main: register staging sstables without throttling
In previous code, registering staging sstable involved taking
a semaphore, but at the same time a bug prevented it from hanging -
futures were just fired away without waiting for them, so the semaphore
was just a little overloaded with useless dangling futures.
To fix that, registering staging sstables is now waited for,
and it's performed without using the throttling semaphore (by passing
a throttle::no flag).