db,view: wait for all hints before marking view as built
In order not to mark view as built prematurely, the implementation
will first wait for all existing view hints for the destination
endpoint to be sent.
Streaming code checks if view building is finished and generates
view updates out of streamed sstables based on this information.
With this change, unsent hints will prevent the view build from
being marked as finished, which may in turn cause more sstables
to be used for view updates generation, which, in turn,
reduces the chance of inconsistencies.

Fixes #4039