cql3: handle map value as index target
In order to support indexes created on a specific map entry (e.g. m[1]),
index target definition is broadened. It used to be a column definition
and now it can be either that or a column definition + map key.
Passing a map value is not yet supported in CQL syntax, it will be
activated after the feature is fully implemented.