view: handle view row liveness with computed columns
In order to properly manage liveness of view rows, additional
information about a computed column present in view primary key
is added - we need to know if there's a regular base column that
is also used to compute a computed column for view primary key.
This information is later used to compute appropriate row markers.
The mechanism is similar to handling regular base column in view pk.