types: add deserializing single item from a collection
Right now the only way to access a single item from a collection
is to deserialize it all into a contiguous buffer.
For map value indexing sake, it makes sense to return only
the item needed, which can save allocating a large buffer
and time spent on deserializing unneeded elements.