database: prevent clogging execution stage by view updates
Applying data to the database is done through a call to
database::apply(), which uses an execution stage based on a helper
do_apply() function, which works under an execution stage.
However, do_apply() can also recursively call apply() if it needs
to apply view updates. That can clog the execution stage in the
following scenario:
1. Execution stage queue is full, do_apply is called.
2. Execution stage queue flush is forced, which results.
   in popping 1 element from the queue calling do_apply() immediately.
3. do_apply calls apply() because of view updates, is immediately
   served because of futures being ready and ends up in do_apply.
4. do_apply is called and the queue is full again - flush is forced.

And so on - instead of flushing as many elements as possible from the
queue, they are popped one by one with considerable overhead every time,
which defeats the purpose of SEDA.

The patch forces scheduling the view-related apply() (3.) for later,
thus removing the immediate recursive call.

Observed result is that execution stage queue does not oscilate between
max_queue_length and max_queue_length -1 anymore.

Refs #4856