alternator: use keys from system_auth.roles for authorization
Instead of having a hardcoded secret key, the server now verifies
an actual key extracted from system_auth.roles system table.
This commit comes with a test update - instead of 'whatever':'whatever',
the credentials used for a local run are 'cassandra':'', which matches
the initial contents of system_auth.roles table, which acts as a key
store.

Fixes #5046