db: allow overriding view update elision
Mutations for the base table undergo additional tests when generating
view updates from them - if a view update is not necessary from the
perspective of liveness and correctness, it will be skipped for
performance benefits.
For view regeneration, an option to override these checks is needed
in order to enforce generating a view update, even if from the base
table's perspective everything looks up to date - that's because
the view row may be missing.
This commit adds a parameter for forcing view update generation
and pipes it through all the interfaces.