cql3: add a statement for regenerating view rows
In order to expose the API for regenerating view rows from their base,
a CQL statement is created. It is loosely based on select_statement,
as its first step is to select base table rows.
Note that beside having high level CQL API, this statement is rather low
level itself: it is going to push view *replica* updates,
which in particular assumes that this node *is* a replica for given
base key and the repair will occur only between this node and its
paired view replica. Thus, this statement is considered a specialized
low level tool, allowing to build repair scripts on top of it.