alternator: replace throws with returns in executor
In order to minimize the usage of throws and catches,
executor.cc is reworked to be based on returning errors instead.
Most apis are changed from:
future<X> f() {
	(...)
	throw api_error(err);
}
to:
future<variant<X,  api_error>() {
	(...)
	return make_ready_future<variant<X, api_error>>(err);
}

Note: other source files are not yet reworked. They still work properly,
because exceptions are still catched and translated in the upper layer,
but they should be rewritten to use std::variant as well.
That includes at least conditions.cc and serialization.cc.

Refs #5472