db,view: introduce sync_local_view_updates parameter
This parameter can be used to enforce synchronous view updates,
if the code determines that the update is going to be local, e.g.
it will be applied locally without contacting other nodes.
That's expected to be the case especially for local indexes
and local views, i.e. views where the partition key definition
is the same as base's.
Currently all code still uses sync_local_view_updates::no,
which effectively always sends view updates asynchronously.