alternator: implement json parser service
The service runs in a static thread which accepts and parses
documents. It is expected to be instantiated by a service
which needs JSON parsing to run in a thread, but at the same time
only allocate the stack for the thread once, not per-request.
Documents smaller than a parsing threshold (currently: 64KiB)
will be parsed in place without yielding. The assumption is that
most alternator requests are small and there's no need to parse
them in a yieldable way, which also induces overhead.
For reference, parsing a 128KiB document with rapidjson takes
around 0.5 millisecond.