messaging_service: introduce isolation_provider
We want to be able to isolate different statement operations, to be used
by query classification. More specifically we want to send statement
verbs on different connections based on the scheduling group they are
run in. We also want to transfer this scheduling group to the server, so
the verb body is also executed in its context.
This is very similar to the requirements of the enterprise service level
isolation. To allow the two to use the same infrastructure this patch
introduces an interface which abstracts away how operations executed in
the context of different scheduling groups are isolated on the
connection level, and how this isolation is transferred to the remote.
Using this interface, we can just have different implementations in the
open-source and enterprise versions. Although this abstraction is not
required as we won't switch between two implementations in the same
binary in the foreseeable future and the enterprise versions can just
adjust I think there is value in moving towards an architecture
extensible infrastructure, where enterprise features are just different
implementations of certain interfaces.

The interface introduced in this patch is called `isolation_provider`
and has following two methods:
* connection_config get_connection_config(default_client_idx client_idx)
* rpc::isolation_config get_isolation_config(sstring isolation_cookie)

The former is called on the client and is used to select the actual
network connection for the logical connection, as well as specify an
isolation configuration. This isolation configuration is then sent to
the server in the form of the isolation_cookie, and mapped back to
actual isolation configuration via get_isolation_config().

This patch also introduces an implementation `no_isolation_provider()`,
which implements the current status quo: all verbs are dispatched on the
default connection, no isolation is provided. For now we pass this to
the messaging service constructor. In the next patch we will introduce
an isolation provider which implements the isolation required for query
classification and will switch migration manager to use it.