messaging_service: introduce statement_classifying_isolation_provider
The statement classifying isolation provider is an isolation provider
implementation which implements the isolation required for query
classification. Operations are classified as either user or system,
based on the scheduling group they run in:
* statement scheduling group -> user
* everything else -> system
This only applies to verbs that use one of the two statement clients
(`default_client_idx::statement0` and `default_client_idx::statement1`).

On client, we send an isolation cookie containing `$system` for
operations classified as system. We send an empty isolation cookie for
all other cases. On the server we use the default isolation
configuration for the verb for non-statement verbs and user statements.
For system statements, we use the configured system scheduling group
(the default scheduling group).

Also switch the migration manager to use this isolation provider.