Add .parquet storage
Proof of concept.

The writer doesn't understand collections (they are written as blobs), counter
columns (they are omitted), and range tombstones yet. Also, some abstract types
are written to parquet as blobs, when they could be written with parquet's
native logical types (e.g. uuid).

I'm not fully confident in the correctness of this solution. I have only tested
it by hand, inserting some values with cqlsh and looking at the outputs.
I'm not sure how to write tests for this.

types.hh:239 says
> Cassandra has a notion of empty values even for scalars (i.e. int).  This is
> distinct from NULL which means deleted or never set.
Is this the case for mc too? I don't understand when and how can that occur.
Anyway, if they happen (or any other deserialization error occurs),
then a NULL will be written to parquet. That can produce an invalid
parquet file, if this happens in partition key.

Caution: the writer doesn't flush pages from memory to disk until it's closed.
This needs to be fixed before testing on big files. I will fix it over the
weekend.