cql3: fix potential segfault on disengaged optional
In untyped_result_set::get_view, there exists a silent assumption
that the underlying data, which is an optional, to always be engaged.
In case the value happens to be disengaged it may lead to creating
an incorrect bytes view from a disengaged optional.
In order to make the code safer (since values parsed by this code
often come from the network and can contain virtually anything)
a segfault is replaced with an exception, by calling optional's
value() function, which throws when called on disengaged optionals.

Fixes #6915
Message-Id: <090d3f323f693eaa833cad551a044a934b06678d.1595602089.git.sarna@scylladb.com>