service: add metrics for successful range read latency
In order to distinguish between operations that succeeded and ones
that returned an error, a new histogram is introduced for successful
range reads only. The new histogram is resilient to situations when
failing requests poison the "range read latency histogram", showing
latencies higher than the timeout - so that it's still possible
to see what's the p99.9 latency for range reads that actually succeeded.