exceptions: add more constructors to overloaded_exception
Historically, overloaded_exception was used only to let the user know
that the number of in-flight hints is too high. Given that "overloaded
exception" is quite a generic name, some more constructors are added
to it, since it can be used in later patches for more overloading
scenarios - e.g. when a number of concurrent reads is too high.